requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
docutils==0.20.1
fonttools==4.43.1
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.3.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
mplfinance==0.12.10b0
nh3==0.2.14
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pkginfo==1.9.6
PlvLogger==0.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
six==1.16.0
twine==4.0.2
tzdata==2023.3
urllib3==2.0.7
websocket-client==1.6.3
zipp==3.17.0
